Update histiry local

+++ lib/views/product/product_special_view.dart
@@ -44,7 +44,7 @@ class _ProductSpecialState extends State<ProductSpecial> {
         id = valueId;
         searchsCompanies(id, widget.lstProduct, productCompanies);
       });
-    }) as int;
+    });
   }
 
   @override
